+++ NEWS.md
@@ -1,6 +1,8 @@
 # ggplot2 (development version)
 
 * Added `panel.widths` and `panel.heights` to `theme()` (#5338, @teunbrand).
+* The `summary()` method for ggplots is now more terse about facets 
+  (@teunbrand, #5989).
 * `guide_bins()`, `guide_colourbar()` and `guide_coloursteps()` gain an `angle`
   argument to overrule theme settings, similar to `guide_axis(angle)` 
   (@teunbrand, #4594).